+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Do not retry incoming activity if access to referenced object is denied.
+
 ## [2.17.1] - 2024-04-29
 
 ### Changed