+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Do not retry incoming activity if access to referenced object is denied.
 
+### Fixed
+
+- Fixed content encoding errors when serving media files.
+
 ## [2.17.1] - 2024-04-29
 
 ### Changed