+++ CHANGELOG.md
@@ -26,6 +26,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fixed key ID to actor ID transformation for microblog.pub.
 - Do not discard object if emoji ID is not valid.
 
+### Security
+
+- Re-fetch announced FEP-1b12 activities.
+
 ## [2.17.1] - 2024-04-29
 
 ### Changed