+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Allow `fetch-object-as` command to be called without username argument.
 - Display better error message if fetched object doesn't have an ID.
 - Set `Relationship.muting_notifications` flag to True if account is muted.
+- Add `blocking` and `blocked_by` properties to Mastodon API Relationship object.
 
 ### Deprecated