+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [2.18.0] - 2024-05-07
+
 ### Added
 
 - Added `/api/v1/settings/import_followers` API endpoint (replaces `/move_followers`).